requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sentence-transformers
langchain
langchain-google-genai
streamlit
python-dotenv
pandas
pydantic
streamlit
chromadb
google-genai
python-dotenv
selenium
webdriver-manager
beautifulsoup4